pyproject.toml
[tool.poetry]
name = "stocks-bot"
version = "0.1.0"
description = "Telegram Bot to track stocks data"
authors = ["Yoav Navon <yanavon@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.4.1"
python-dotenv = "^0.17.0"
yfinance = "^0.1.59"
matplotlib = "^3.4.1"
seaborn = "^0.11.1"
boto3 = "^1.17.60"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
